Moving metrics into results.py
